Initial commit
Starting with Alexander Jung's (ajdotnet) work, I ripped out the console
portion of the solution and tweaked the Set-DisplaySettings cmdlet to
allow an unreported resolution to be forced. This is useful because
custom resolutions created in the NVidia control panel don't show up as
supported but it's still possible to use them.